minor version number cleanup
those version numbers look like salt version numbers and not idem ones. The last bit about the output of 'idem exec test.version' is invented. It blows up on my site so I do not know what this would actually report? I assume it might be the output in invented, but maybe that is well off.